chore: upgrade github client library
